# backend/openapi.yaml
openapi: 3.1.0
info:
  title: LangGraph Chatbot API
  description: AI-powered chatbot with conversation history using LangGraph and MongoDB
  version: 1.0.0
paths:
...
  /chatbot/chat:
    post:
      tags:
      - Chatbot
      summary: Chat Endpoint
      description: "Send a message to the chatbot and receive a response.\n\nProcesses\
        \ user message through LangGraph workflow, saves conversation to MongoDB,\n\
        tracks token usage, and updates chat metadata.\n\nArgs:\n    chat_request:\
        \ Message and optional chat_id and model_name\n    current_user: Authenticated\
        \ user from JWT token\n\nReturns:\n    Assistant's response, chat_id, and\
        \ usage metadata"
      operationId: chat_endpoint_chatbot_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /chatbot/chats:
    get:
      tags:
      - Chatbot
      summary: Get All Chats
      description: "Retrieve all chat sessions for the authenticated user.\n\nReturns\
        \ chat metadata sorted by most recently updated.\n\nArgs:\n    current_user:\
        \ Authenticated user from JWT token\n\nReturns:\n    List of chat sessions\
        \ and total count"
      operationId: get_all_chats_chatbot_chats_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatListResponse'
      security:
      - OAuth2PasswordBearer: []
  /chatbot/chats/{chat_id}:
    get:
      tags:
      - Chatbot
      summary: Get Specific Chat
      description: "Retrieve a specific chat session with full message history.\n\n\
        Fetches chat metadata and all messages, formatting them with role, content,\n\
        timestamp, and usage metadata.\n\nArgs:\n    chat_id: Unique identifier for\
        \ the chat session\n    current_user: Authenticated user from JWT token\n\n\
        Returns:\n    Chat metadata and formatted message history\n\nRaises:\n   \
        \ HTTPException: If chat not found or doesn't belong to user"
      operationId: get_specific_chat_chatbot_chats__chat_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          title: Chat Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Chatbot
      summary: Delete Chat
      description: "Delete a specific chat session and all associated messages.\n\n\
        Removes chat metadata and messages from MongoDB. Admin messages are soft-deleted\n\
        (marked with deleted: true) for audit purposes.\n\nArgs:\n    chat_id: Unique\
        \ identifier for the chat session\n    current_user: Authenticated user from\
        \ JWT token\n\nReturns:\n    Success message with deleted chat_id\n\nRaises:\n\
        \    HTTPException: If chat not found or doesn't belong to user"
      operationId: delete_chat_chatbot_chats__chat_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          title: Chat Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteChatResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /token-usage/summary:
    get:
      tags:
      - Token Usage
      summary: Get Token Usage Summary
      description: Get aggregated token usage statistics for the authenticated user.
      operationId: get_token_usage_summary_token_usage_summary_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: model_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by specific LLM model
          title: Model Name
        description: Filter by specific LLM model
      - name: days
        in: query
        required: false
        schema:
          type: integer
          description: Number of days to look back
          default: 30
          title: Days
        description: Number of days to look back
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenUsageSummaryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /token-usage/history:
    get:
      tags:
      - Token Usage
      summary: Get Token Usage History
      description: Get detailed token usage history with pagination.
      operationId: get_token_usage_history_token_usage_history_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: model_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by specific LLM model
          title: Model Name
        description: Filter by specific LLM model
      - name: chat_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by specific chat session
          title: Chat Id
        description: Filter by specific chat session
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          description: Maximum number of records to return
          default: 100
          title: Limit
        description: Maximum number of records to return
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          description: Number of records to skip
          default: 0
          title: Skip
        description: Number of records to skip
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenUsageHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /token-usage/by-chat/{chat_id}:
    get:
      tags:
      - Token Usage
      summary: Get Chat Token Usage
      description: Get aggregated token usage for a specific chat session.
      operationId: get_chat_token_usage_token_usage_by_chat__chat_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: chat_id
        in: path
        required: true
        schema:
          type: string
          title: Chat Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatTokenUsageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /token-usage/models:
    get:
      tags:
      - Token Usage
      summary: Get Used Models
      description: Get list of all LLM models used by the authenticated user.
      operationId: get_used_models_token_usage_models_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsedModelsResponse'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    ChatDetailResponse:
      properties:
        chat_metadata:
          $ref: '#/components/schemas/ChatMetadata'
        messages:
          items:
            $ref: '#/components/schemas/MessageItem'
          type: array
          title: Messages
      type: object
      required:
      - chat_metadata
      - messages
      title: ChatDetailResponse
      description: Response for a specific chat with messages.
    ChatListResponse:
      properties:
        chats:
          items:
            $ref: '#/components/schemas/ChatMetadata'
          type: array
          title: Chats
        total:
          type: integer
          title: Total
      type: object
      required:
      - chats
      - total
      title: ChatListResponse
      description: Response for listing all chats.
    ChatMetadata:
      properties:
        SessionId:
          type: string
          title: Sessionid
        user_id:
          type: string
          title: User Id
        chat_id:
          type: string
          title: Chat Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        message_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Message Count
        total_input_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total Input Tokens
        total_output_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total Output Tokens
        total_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total Tokens
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      additionalProperties: true
      type: object
      required:
      - SessionId
      - user_id
      - chat_id
      title: ChatMetadata
      description: Metadata for a chat session.
    ChatModelUsage:
      properties:
        model_name:
          type: string
          title: Model Name
        total_input_tokens:
          type: integer
          title: Total Input Tokens
        total_output_tokens:
          type: integer
          title: Total Output Tokens
        total_tokens:
          type: integer
          title: Total Tokens
        request_count:
          type: integer
          title: Request Count
      type: object
      required:
      - model_name
      - total_input_tokens
      - total_output_tokens
      - total_tokens
      - request_count
      title: ChatModelUsage
      description: Token usage for a single model within a specific chat session.
    ChatRequest:
      properties:
        message:
          type: string
          title: Message
        chat_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Chat Id
        model_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Name
      type: object
      required:
      - message
      title: ChatRequest
    ChatResponse:
      properties:
        response:
          type: string
          title: Response
        chat_id:
          type: string
          title: Chat Id
        usage:
          anyOf:
          - $ref: '#/components/schemas/UsageMetadata'
          - type: 'null'
      type: object
      required:
      - response
      - chat_id
      title: ChatResponse
      description: Response from the chatbot for a single message.
    ChatTokenUsageResponse:
      properties:
        user_id:
          type: string
          title: User Id
        chat_id:
          type: string
          title: Chat Id
        models:
          items:
            $ref: '#/components/schemas/ChatModelUsage'
          type: array
          title: Models
        totals:
          $ref: '#/components/schemas/UsageTotals'
      type: object
      required:
      - user_id
      - chat_id
      - models
      - totals
      title: ChatTokenUsageResponse
      description: Aggregated token usage for a specific chat session.
...
    DeleteChatResponse:
      properties:
        message:
          type: string
          title: Message
        chat_id:
          type: string
          title: Chat Id
      type: object
      required:
      - message
      - chat_id
      title: DeleteChatResponse
      description: Response confirming chat deletion.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageItem:
      properties:
        role:
          type: string
          title: Role
          description: Either 'user' or 'assistant'
        content:
          type: string
          title: Content
        timestamp:
          anyOf:
          - type: string
          - type: 'null'
          title: Timestamp
        usage_metadata:
          anyOf:
          - $ref: '#/components/schemas/UsageMetadata'
          - type: 'null'
          description: Token usage metadata, only present on assistant messages
      type: object
      required:
      - role
      - content
      title: MessageItem
      description: A single message item in a chat conversation.
...
    ModelInfo:
      properties:
        model_name:
          type: string
          title: Model Name
        usage_count:
          type: integer
          title: Usage Count
        last_used:
          type: string
          format: date-time
          title: Last Used
      type: object
      required:
      - model_name
      - usage_count
      - last_used
      title: ModelInfo
      description: Brief info about a model used by the current user.
    ModelUsage:
      properties:
        model_name:
          type: string
          title: Model Name
        total_input_tokens:
          type: integer
          title: Total Input Tokens
        total_output_tokens:
          type: integer
          title: Total Output Tokens
        total_tokens:
          type: integer
          title: Total Tokens
        request_count:
          type: integer
          title: Request Count
        first_used:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: First Used
        last_used:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used
      type: object
      required:
      - model_name
      - total_input_tokens
      - total_output_tokens
      - total_tokens
      - request_count
      title: ModelUsage
      description: Token usage statistics for a single model over a given period.
    OutputTokenDetails:
      properties:
        reasoning:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reasoning
        audio:
          anyOf:
          - type: integer
          - type: 'null'
          title: Audio
        text:
          anyOf:
          - type: integer
          - type: 'null'
          title: Text
      additionalProperties: true
      type: object
      title: OutputTokenDetails
      description: Granular breakdown of output tokens, as reported by the LLM provider.
...
    TokenUsageHistoryResponse:
      properties:
        user_id:
          type: string
          title: User Id
        total_records:
          type: integer
          title: Total Records
        returned_records:
          type: integer
          title: Returned Records
        skip:
          type: integer
          title: Skip
        limit:
          type: integer
          title: Limit
        records:
          items:
            $ref: '#/components/schemas/TokenUsageRecord'
          type: array
          title: Records
      type: object
      required:
      - user_id
      - total_records
      - returned_records
      - skip
      - limit
      - records
      title: TokenUsageHistoryResponse
      description: Paginated token usage history for a user.
    TokenUsageRecord:
      properties:
        user_id:
          type: string
          title: User Id
        chat_id:
          type: string
          title: Chat Id
        model_name:
          type: string
          title: Model Name
        input_tokens:
          type: integer
          title: Input Tokens
        output_tokens:
          type: integer
          title: Output Tokens
        total_tokens:
          type: integer
          title: Total Tokens
        input_token_details:
          anyOf:
          - $ref: '#/components/schemas/InputTokenDetails'
          - type: 'null'
        output_token_details:
          anyOf:
          - $ref: '#/components/schemas/OutputTokenDetails'
          - type: 'null'
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
      - user_id
      - chat_id
      - model_name
      - input_tokens
      - output_tokens
      - total_tokens
      - timestamp
      title: TokenUsageRecord
      description: A single token usage record for one LLM request.
    TokenUsageSummaryResponse:
      properties:
        user_id:
          type: string
          title: User Id
        period_days:
          type: integer
          title: Period Days
        start_date:
          type: string
          format: date-time
          title: Start Date
        end_date:
          type: string
          format: date-time
          title: End Date
        filter_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Filter Model
        models:
          items:
            $ref: '#/components/schemas/ModelUsage'
          type: array
          title: Models
        totals:
          $ref: '#/components/schemas/UsageTotals'
      type: object
      required:
      - user_id
      - period_days
      - start_date
      - end_date
      - filter_model
      - models
      - totals
      title: TokenUsageSummaryResponse
      description: Aggregated token usage summary for a user over a date range.
    UsageMetadata:
      properties:
        input_tokens:
          type: integer
          title: Input Tokens
        output_tokens:
          type: integer
          title: Output Tokens
        total_tokens:
          type: integer
          title: Total Tokens
        input_token_details:
          anyOf:
          - $ref: '#/components/schemas/InputTokenDetails'
          - type: 'null'
        output_token_details:
          anyOf:
          - $ref: '#/components/schemas/OutputTokenDetails'
          - type: 'null'
      type: object
      required:
      - input_tokens
      - output_tokens
      - total_tokens
      title: UsageMetadata
      description: Token usage metadata returned by the LLM for a single request.
    UsageTotals:
      properties:
        total_input_tokens:
          type: integer
          title: Total Input Tokens
        total_output_tokens:
          type: integer
          title: Total Output Tokens
        total_tokens:
          type: integer
          title: Total Tokens
        total_requests:
          type: integer
          title: Total Requests
      type: object
      required:
      - total_input_tokens
      - total_output_tokens
      - total_tokens
      - total_requests
      title: UsageTotals
      description: Aggregated token usage totals across all models.
    UsedModelsResponse:
      properties:
        user_id:
          type: string
          title: User Id
        models:
          items:
            $ref: '#/components/schemas/ModelInfo'
          type: array
          title: Models
      type: object
      required:
      - user_id
      - models
      title: UsedModelsResponse
      description: List of all LLM models the user has used.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/login/user